v0.0.4
Adds missing "plus.rectangle" symbol to controller symbol collection. Fixes issue where the ps4s touchpad button would not display a glyph. 
Added method SupportsSFSymbols to GCControllerElement to detect if we are in iOS14. Works for iOS, but not for macOS yet. 
Added fallback mechanism for SFSymbols on iOS13
Rewired controller elements are now mapped to apple controller elements using an asset, instead of being hardcoded. Fixes an issue where the elementID of the custom controllers elements would be different in different projects, depending on the order in which the elements were created.